script: Lenovo Legion Go LCD display configuration
Add support for the Lenovo Legion Go handheld, which features a rotated
1600x2560 panel that reports 60Hz and 144Hz modes in the EDID. VRR and
HDR are not supported, and only one panel model is known to be in use.

Modes other than 60Hz and 144Hz have a small chance of causing a locked
touchscreen until another modeset is performed or have artifacts during
modesetting, so extra modes are not included. The dynamic_modegen section
is filled out in case users decide to add their own refresh rates in a
local table.

This configuration has been tested with:
* SteamOS Main 20241025.1000 with kernel 6.8.12-valve3
* Arch Linux with kernel 6.12-rc5

Signed-off-by: Matthew Schwartz <matthew.schwartz@linux.dev>