fix: enable HDR colorspace only when requested
Currently, when using a PQ HDR10 display, if the user enables HDR, the
output colorspace is always PQ, which causes additional battery drain,
brightness quirks, and can potentially cause artifacting on Nvidia.

Instead of enabling it always, detect when an app requests HDR and
only enable it afterwards. Retain HDR capability until the app switches
to SDR or the user exits it.