fix: enable HDR colorspace only when requested
Currently, when using a PQ HDR10 display, if the user enables HDR for
externals or always for internals, the output colorspace is always PQ,
which causes additional battery drain, brightness quirks, and can
potentially cause artifacting on Nvidia.

Move HDR advertisment to occur based on support. Then, instead of
enabling HDR always, detect when an app requests HDR and only enable it
afterwards. Retain HDR capability until the app switches to SDR or the
user exits it.